refactor: components 폴더 정리
