+++ src/pages/LandingPage/sections/Carousel.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { css } from '@emotion/react';
-import LeftArrowButton from '../assets/svg/LeftArrowButton';
-import RightArrowButton from '../assets/svg/RightArrowButton';
-import Pagination from './Pagination';
+import LeftArrowButton from '../../../assets/svg/LeftArrowButton';
+import RightArrowButton from '../../../assets/svg/RightArrowButton';
+import Pagination from '../../../components/common/Pagination';
 
 type CarouselProps = {
   items: JSX.Element[];
