+++ src/pages/MyPage/HistoryPage/HistoryPage.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
-import Tab from '../../../components/Mypage/TopTab/Tab/Tab';
-import TopTab from '../../../components/Mypage/TopTab/TopTab';
 import { PATH } from '../../../constants/path';
 import { historyContainer } from './HistoryPage.style';
+import Tab from '../MypageSections/TopTab/Tab/Tab';
+import TopTab from '../MypageSections/TopTab/TopTab';
 
 const HistoryPage = () => {
   const tabs = [
