+++ src/pages/PostListPage/PostListPage.tsx
@@ -3,8 +3,8 @@ import { css } from '@emotion/react';
 import { useQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import PostItem from './sections/PostItem';
-import SortButton from '../../components/SortButton';
-import ToggleButton from '../../components/ToggleButton';
+import SortButton from '../../components/Buttons/SortButton';
+import ToggleButton from '../../components/Buttons/ToggleButton';
 import { fetchPostsData } from '../../api/posts/posts';
 
 const PostList = () => {
